requirements.txt
flask
numpy
pandas
Flask-SocketIO
selenium
beautifulsoup4
python-socketio[client]